restructure as suggested by @tardieu
