+++ tests/_inductor/utils_inductor.pyi
